radio_receiver/requirements.txt
pyrtlsdr
scipy
numpy
pyaudio
SpeechRecognition